fuse: Fix narrowing error in switch statement

+++ src/client/fuse_ll.cc
@@ -608,7 +608,7 @@ static void fuse_ll_flush(fuse_req_t req, fuse_ino_t ino,
 
 #ifdef FUSE_IOCTL_COMPAT
 static void fuse_ll_ioctl(fuse_req_t req, fuse_ino_t ino, int cmd, void *arg, struct fuse_file_info *fi,
-                          unsigned flags, const void *in_buf, size_t in_bufsz, size_t out_bufsz)
+			  unsigned flags, const void *in_buf, size_t in_bufsz, size_t out_bufsz)
 {
   CephFuse::Handle *cfuse = fuse_ll_req_prepare(req);
 
@@ -617,7 +617,7 @@ static void fuse_ll_ioctl(fuse_req_t req, fuse_ino_t ino, int cmd, void *arg, st
     return;
   }
 
-  switch(cmd) {
+  switch (static_cast<unsigned>(cmd)) {
     case CEPH_IOC_GET_LAYOUT: {
       file_layout_t layout;
       struct ceph_ioctl_layout l;
